Fixes ptrUid lost when using impl types with QI_REGISTER_IMPLEMENTATION_H
QI_REGISTER_IMPLEMENTATION_H associates an interface type to an
implementation type that might be unrelated. However,
not like QI_REGISTER_IMPLEMENTATION (note the absence of `_H`),
QI_REGISTER_IMPLEMENTATION_H generates special proxy types wrapping
the real implementation object.

When using QI_REGISTER_IMPLEMENTATION_H and doing:

    Object<Interface> obj{ make_shared<Impl> }; // Will instantiate wrapper

then the Object constructor detects this situation as a special case
and instantiates the appropriate wrapper type then put the real
implementation object in it.
In the process, the ptrUid of the real objet is lost (without this patch)
when creating the GenericObject associated to this wrapper+impl couple,
but it is still the real object being called in the end.

In this change we are making sure that we provide the real implementation
ptrUId instead of creating a new one based on the wrapper object
when making the associated GenericObject.

Change-Id: Iedca59079d4d2f199517e332a4fd4bd96a538057
Reviewed-on: http://gerrit.aldebaran.lan/91604
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>