qi.messaging: Makes `ServiceDirectoryProxy` implementation less blocking.
There were a few call to Future.value() in the implementation functions
of the service directory proxy. This patch replaces them with
continuations.

This made the MirroringResult structure useless because errors can
now directly be put in the futures returned by the functions. Instead
we have a MaybeMirroredId that is simply an optional to a integer.

Change-Id: Ia69ad8cdfc9314a1b8ea14954f88155e0a302e91
Reviewed-on: http://gerrit2.aldebaran.lan/32
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>