ka.functional: Adds function and action composition with `|`.
The traditional mathematical order of function composition is represented by
`*` in `ka`: `g * f` is a function applying `f` then `g`.

It is sometimes easier to read composition from left to right. Therefore, `|`
is identical to `*` except that argument order is reversed: `f | g` is a
function that applies `f` then `g`.

This applies to function composition as well as action composition.

Change-Id: I3439200323bfff78be5e59423bc5c91f32763387
Reviewed-on: http://gerrit.aldebaran.lan/92405
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: julien.bernard <julien.bernard@external.softbankrobotics.com>