ka.iterator: Adds operators for iterators.
It is useful to write generic code that accepts indistinctly arithmetic types and
iterator types.

For example, incrementing arithmetic types and iterator types is similar. Same
for the distance between two numbers / iterators. In the case of iterators, the
complexity of these operations depends on the iterator category.

For a formal justification of these operators, see [Elements of Programming
(Stepanov-MacJones, 2009)](http://elementsofprogramming.com/), section 6.3.

Change-Id: Ia538b841bd6ed2f38a1db4b6cde1ca1ee0eb8392
Reviewed-on: http://gerrit.aldebaran.lan/92507
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: julien.bernard <julien.bernard@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>