qi.socket: Forwards the received message as non-const inside socket code.
Due to some protocol extensions (e.g. direct message dispatch), it may now be
needed in `TcpMessageSocket` to modify the received message. Until now, the
received message was internally forwarded as a const reference, so it was not
modifiable by `TcpMessageSocket`.

Note that the message is still passed as a const-reference to upper layers, so
that libqi's code is not impacted.

Change-Id: I7fc3ae6f6e3903167ea2aa6f9105241015305121
Reviewed-on: http://gerrit.aldebaran.lan/92552
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>