tests: Replaces duration subtypes by native integers in exposed functions in test_call.
Duration types are converted to integers representing nanoseconds in
messages. It is incorrect to call a function expecting for example
MilliSeconds by passing it an integer and expecting it to be considered
as milliseconds, as it will be instead be considered as nanoseconds.

Example: Failing to call a method that exposes MilliSeconds argument
         correctly.
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
void foo(qi::MilliSeconds dura);
...
obj.call<void>("foo", 500); // 500 here will not mean 500 milliseconds,
                            // but 500 nanoseconds converted to milliseconds
                            // which is 0 millisecond.
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Change-Id: Ia9788d63dde5c0743adf2a5158d23955b9cde33d
Reviewed-on: http://gerrit.aldebaran.lan/93749
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: julien.bernard <julien.bernard@external.softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>