qi.messaging: Adds tests for message dispatching. #42553
These tests compare the duration of a call to a non-service object,
vs. the same call but when there are a lot of remote objects stored
in the caller process.
The goal is to make obvious the issue of message dispatching duration
taking more time when a lot of `RemoteObject` are persisted in the
process having the socket receiving the message.

Change-Id: I2f4f340a2589d8b42e8eca2a8b160585eeb33135
Reviewed-on: http://gerrit.aldebaran.lan/94368
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>