buffer: Removes memory pool.
The pool was never shrunk, resulting in ever increasing memory. This problem
is particularly salient in case of isolated message bursts.

It may also get worse system-wise due to the existence of independent libqi
instances in distinct processes.

Another potential aggravating factor (not addressed by this patch) is that each
buffer is currently ~768 bytes long at least, which may have a negative impact
on memory in case of a small message burst. This lead can also be explored.

This patch must be tested in real context to prove its value.

This is to mitigate issues #43516 and #43454.

Change-Id: I9e03ec0f3dbc70be7ca1bca14db617ef6775938a
Reviewed-on: http://gerrit.aldebaran.lan/94091
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>