qi.eventloop: Adds a dynamic shrinking mechanism.
General explanation
=====================

Up to now, the behavior of the threadpool was to spawn threads on contention.
These threads were never destroyed. Having a temporary burst of activity
(typically at startup) would later result in a lot of idle threads, causing an
unjustified increase of cpu load. The threadpool now terminates threads that
have been idle for too long.

Environment variables
=====================

The maximum idle duration is controlled by the environment variable
`QI_EVENTLOOP_THREAD_MAX_IDLE_DURATION` (in milliseconds, like the other
eventloop environment variables).

It is also possible to specify a minimum number of threads with the environment
variable `QI_EVENTLOOP_MIN_THREAD_COUNT`. This allows to limit the shrinking
mechanism, even if the threads are idle. A `setMinThreads` method is available
on `Eventloop` too.

Technical details
=====================

The infamous "ping thread" was responsible for spawning new threads in case of
contention. It is now, on top of that, responsible for terminating idle threads.
Each thread is now associated to a last task run date. The "ping task"
(periodically launched by the "ping thread" to check contention) now checks if
this date for the current worker thread is too far in the past. If it is, a
specific exception is launched that causes the termination of this worker
thread. The general behavior of the threadpool is that if it has too many
threads, they are bound to sooner or later run this ping task and therefore to
be terminated.

Change-Id: Ieb9eddbe8c35a87b39b12bd04ae7871a3635d2bb
Reviewed-on: http://gerrit.aldebaran.lan/94133
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>