TypeInfo: Uses boost::type_index instead of std::type_info #43651
On some platforms we can not rely on std::type_info (due to known
implementation limitations). boost::type_index is a reliable replacement.
See the boost documentation for an exact description of the issues: 
https://www.boost.org/doc/libs/1_64_0/doc/html/boost_typeindex.html#boost_typeindex.motivation

Libqi type interface mechanism relied on std::type_info. As a consequence
it was broken on some platforms (for example, on android-arm64 with
the clang toolchain).

Change-Id: I2f43fcb355ea194ad6ab44f042693ccc4ec3e445
Reviewed-on: http://gerrit.aldebaran.lan/94291
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: philippe.martin <philippe.martin@softbankrobotics.com>