Adds ObjectUid as universal ID of remote objects implemented using PtrUid.
This change removes the implication that we are exposing pointers
in qi::Object interface through qi::Object::ptruid().
We deprecate qi::Object::ptruid() and provide qi::Object::uid() as
a clearer replacement.
The implementation still relies on ObjectUid to have the interface
and be compatible with PtrUid but we can now separate the meaning
of an unique identifier of an object refered by qi::Object from an
universally unique pointer value that PtrUid represents,
allowing for ObjectUid to be implemented differently in the future
and to clarify the role of qi::Object::uid().

Change-Id: I3fdc96ad9a545d505b6f7c65f93190c74ed026b6
Reviewed-on: http://gerrit.aldebaran.lan/94337
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>