strand: Changes order of actions when joining.
When joining the strand, we waited for the end of the currently
executing task before setting scheduled and deferred tasks futures in
error.

This could result in a deadlock if a stranded task was waiting for the
result of another stranded task when the strand was joined.

We now set scheduled and deferred tasks futures in error before waiting
for the current task to finish.

This patch also refactors the test
"TestStrand.JoinStrandEndsRunningAsyncFutureInError" because it was
racy and could easily deadlock. In consequence, it got renamed to
"TestStrand.JoinStrandEndsScheduledTaskInError".

Change-Id: I4eb40467a377d7a32bfd901d4ddd4d6edbd9f604
Reviewed-on: http://gerrit.aldebaran.lan/94404
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>