strand: Makes a constructor enabled even without QI_WITH_TESTS.
The constructor is used by other constructors of Strand (by delegation)
and thus it was causing a build error when libqi was built without
QI_WITH_TESTS defined.

This patch adds a define when compiling targets in the tests
subdirectory and makes the constructor public only when the new define
QI_IS_TEST is defined and private when it's not.

Change-Id: Id4b202915a9aa4a923cbe6a64487324512ed9fa2
Reviewed-on: http://gerrit.aldebaran.lan/94409
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>