messaging: Replaces use of `StreamContext*` by `MessageSocketPtr`.
`StreamContext*` was used to access the shared capabilities in
deserialization and as keys for remote references in `ObjectHost`. In a
future commit, `ObjectHost` will be rewritten and will need to handle
the lifetime of the `StreamContext/MessageSocket` objects through the
`shared_ptr`s owning them.

This patch replaces all occurences of `StreamContext*` by
`MessageSocketPtr` in message serialization/deserialization so that when
an object is deserialized, it can be assigned to its host and associated
to a `MessageSocket`.

It also removes the `Message::appendValue` method that was declared but
never defined.

Change-Id: Ib5bdfcb0e47bcbdb34b962812163b920b88500c9
Reviewed-on: http://gerrit.aldebaran.lan/94729
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>