objects: Removes old `BoundObject` and renames `ServiceBoundObject` into `BoundObject`.
`BoundObject`s were always `ServiceBoundObject`s and in some places
in the code, it was assumed to be the case. `BoundObject` was also
empty.

This patch removes the old `BoundObject` type in favor of
the `ServiceBoundObject` class renamed into `BoundObject` and clarifies the name
of some types (such as `BoundAnyObject` which was in fact a `shared_ptr<BoundObject>`
which is now a `shared_ptr<BoundObject>` named `BoundObjectPtr`).

Change-Id: I54b9112767168626c757762d034369a716b9a29c
Reviewed-on: http://gerrit.aldebaran.lan/94730
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>