qi.messaging: Increases `QI_MAX_MESSAGE_PAYLOAD` default value.
This change is motivated by recent needs to transfert bigger `qi::Buffer`
objects.

This is *not* the best way to transfer big data. See `qicore/qi::FilePtr` for
a good alternative.
However, we realise that this limitation is an artificial pessimisation
and is not really useful in practice (nor does it help prevening DOS
attacks).

It is also better to let the user decide what is a reasonable maximum
message size for their use case instead of imposing an artificial limit
in the networking libary.
Therefore, we plan to remove this payload limit in the future.

Change-Id: I6e7fbb585023a7829e845ddc6863c1833d42f323
Signed-off-by: Joël LAMOTTE <jlamotte@softbankrobotics.com>
Reviewed-on: http://gerrit.aldebaran.lan/94430
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>