tools: Rewrites the wireshark dissector for the Qi messaging protocol.
The dissector is now capable of dissecting multiple Qi messages in a
single TCP packet and a Qi message splitted in multiple TCP packets.

It also enhances the way wireshark displays the dissected informations
for example by providing better description of message fields or
by dissecting each bit of the flag bitfield.

Change-Id: Ib168d3a2b9244ca150641bc27f75aab30cdd997c
Reviewed-on: http://gerrit.aldebaran.lan/95042
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: julien.bernard <julien.bernard@external.softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>