qi.messaging: Changes the call to `service` in `waitForService`.
The call to `Session_Service::service` in `Session::waitForServiceImpl`
was unneeded and too heavy for its use.

To construct a service object, we need at least 3 RPC:
  - A call to the ServiceDirectory to get the information of the service
  (endpoints, id, etc).
  - A request to connect/authentify on the service.
  - A call to get the meta object of the service object.

But in the case of `waitForService`, we just need to check if the
service already exists, so we can just ask the service directory and stop
after step 1. Therefore this patch replaces the service object request
to a call to `ServiceDirectoryClient::service`.

Change-Id: I1710a46b096f855f7095c4501b1147f2e2d686e5
Reviewed-on: http://gerrit.aldebaran.lan/94587
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>