uuid: Improves cryptographical quality of our internal uuid generator.
"Quality" here is the "quality of randomness" as described on wikipedia:
https://en.wikipedia.org/wiki/Cryptographically_secure_pseudorandom_number_generator

In Boost 1.64 (which is the currently supported version),
the default Boost.UUID random generator uses a Mersenne Twister generator
whose randomness is not considered "cryptographically secure".
This change replaces that random generator by Boost.Random's
`random_device` which is guaranteed (per the documentation) to be a
cryptographically secure random source (of the system).


Change-Id: I007191f1cb7c4aa05c8efaa5a3dd06e415947f0e
Reviewed-on: http://gerrit.aldebaran.lan/94746
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>