qi.messaging: Replaces service uid by a string representation. #94837
Also adds string serialization for ObjectUid.

This is a retro-compatibility fix:
Libqi tries to read the type of a field that should be ignored in case
of struct extension, which is a bug. If the type is not known, it
incorrectly triggers an error (as the field should not have been
observed).
Unfortunately we cannot easily upgrade all user code, so for now we will
use a string as it is a type known by all libqi versions.

Change-Id: I142947946acfb1aa47f808f73e24fc8b351d2693
Reviewed-on: http://gerrit.aldebaran.lan/94837
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>