ka: Adds a unit type and the associated constant.
`unit_t` is a regular type that conveys no information.
`unit` is the only value of this type.

It can be used in place of `void`, with the benefit of acting as a regular
type (storable in data structures, comparable, etc.), and thus avoiding
`void` specialization in generic code.

Theoretically, it corresponds to the terminal object in the category of
sets and functions, which is the empty set. It also happens to be
(isomorphic to) the product of no set, hence the definition as an
empty tuple.

See <https://en.wikipedia.org/wiki/Unit_type>

Change-Id: I2d9ce7a09a4edf60c404fb954fe9a794edca33c1
Reviewed-on: http://gerrit.aldebaran.lan/94886
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>