appsession: Fixes returning a dangling reference.
`ApplicationSession::Config::defaultUrl` returns a reference and it was
returning the result of `connectUrl().value()` which was a reference to
a local variable. We want to avoid changing the return type to keep the
API compatibility.

This patch makes `connectUrl` return a const ref to the optional so that
the result of `.value()` is not a dangling ref anymore.

Change-Id: Iab1fa63b149f16bfb285265141117e3c019798db
Reviewed-on: http://gerrit.aldebaran.lan/94931
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: sbarthelemy <sbarthelemy@aldebaran.com>
Reviewed-by: julien.bernard <julien.bernard@external.softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>