appsession: Fixes incorrect impl of configuration creation.
Fixes #43902, #43895 and partially #43903.

The implementation of `finalizeConfig` was incorrect and would sometimes
fail to complete URLs depending on the default config/environment/CLI args
combination of parameters.

This patch reworks the implementation of the method, declares it in new
header/source so that it can be tested more easily and renames it
`reconfigureWithProgramOptions`.

It adds a lot more tests for it and removes
"test_applicationsession_opts" tests that were less complete and harder
to launch and automatize.

It refactors how the environment variables for `ApplicationSession` are
parsed by using `boost::program_options`.

Change-Id: Ic69e1e0f827bd49cf6c3ceefac96cd1f75275dd5
Reviewed-on: http://gerrit.aldebaran.lan/94932
Reviewed-by: sbarthelemy <sbarthelemy@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>