os.android: Disables use of SD card as home directory.
Partially fixes #40532.
Partially fixes #42942.

The path "/mnt/sdcard" is not guaranteed to be accessible and even less
writable from libqi on Android, therefore it is a bad candidate for the
home directory. There does not seem to be another accessible candidate
from libqi without having the Android context of the application that
embeds libqi, so we make the call fail (with an exception).

If users need to get some path on Android, they must either use the
Android API or the NDK.

Also rewrites the error message of both `home` and `gethostname`
to be clearer.

Change-Id: I7175162c8e120a817ba14d9279343b476cb47c07
Reviewed-on: http://gerrit2.aldebaran.lan/218
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: gerrit <gerrit@aldebaran-robotics.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>