messaging: Fixes a deadlock between ObjectHost and BoundObject.
In some cases, a deadlock could occur between `ObjectHost::removeObject`
and `BoundObject::unbindFromSocket`. The first function destroys the
SocketBindings of a BoundObject which calls the second function, which
then calls `ObjectHost::removeObjectsFromSocket`.

Both the first and the last functions share access to a
`synchronized_value`. This patch fixes it by unlocking the
`synchronized_value` in `ObjectHost` before destroying the BoundObjects
SocketBindings.

Change-Id: Ia7da928ec8427b16ca454eade99a92cce0d362ff
Reviewed-on: http://gerrit.aldebaran.lan/95974
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: julien.bernard <julien.bernard@external.softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>