qi.messaging: Tracks `BoundObject` lifetime from its `weak_ptr`.
`BoundObject` now inherits from `boost::enable_shared_from_this` and we
rely on that to track its lifetime, instead of a `Trackable` member.

This avoids the deadlock that could occur in the destructor of
`BoundObject` when we would wait for the tracked to finish from inside
a tracked callback.

As a result, the code that was added to handle this deadlock becomes
irrelevant, and is therefore removed.

Change-Id: Ie236138fa4016c827090102d54ca7b725ec7a8ce
Reviewed-on: http://gerrit.aldebaran.lan/96214
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>