qi.messaging: Fixes deadlock between `BoundObject` and `ObjectHost`.
There is a circular dependency between `ObjectHost` and `BoundObject`.

When a `BoundObject` object receives a `terminate` message, it tries to
remove itself from its owner (an `ObjectHost` object). This sequence
locks, in order, a mutex in the `BoundObject` that protects its internal
data and then a mutex in `ObjectHost` that protects the object/socket
bindings.

So the thread T1 that executes this sequence locks two mutexes M1 then
M2.

When a socket disconnects from a libqi TCP server, the `Server`
object tries to remove all bound objects related to this socket from the
service bound objects it contains. This sequence locks in order the mutex
in each of service bound objects as `ObjectHost` that protects the
object/socket bindings and then a mutex in each of the `BoundObject`
objects that protects its internal data.

The thread T2 that executes this sequence locks two mutexes M2 then
M1.

╔════╗         ╔════╗
║ T1 ║         ║ T2 ║
╚════╝         ╚════╝
  ▼              ▼
  |              |
  locks M1       |
  ▼              locks M2
  |              ▼
  blocks on M2   |
  ▼              blocks on M1
  |              ▼
threads block one another = deadlock

Obviously this can cause deadlocks if operations are not executed
carefully and if mutexes are not released early enough. Therefore we need
to be careful when modifying methods of this class.

This patch makes sure that bindings in `ObjectHost` are destroyed
outside of the lock that protects the bindings storage.

Change-Id: I5f3ea73968e829170cabeb488ef5c5109c283c34
Reviewed-on: http://gerrit.aldebaran.lan/96217
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>