qi.messaging: Makes socket cache return a better error message.
Fixes #42944.

When no endpoint was acceptable to the `TransportSocketCache`, the
`ConnectionAttempt` would end up being destroyed without being added to
the map of attempts, which resulted in a "Promise broken" error back to
the user, that could occur for example when the user used
`Session::service`.

This patch adds a destructor to `ConnectionAttempt` so that if the
promise is still running, an error is set with a better message than
just "Promise broken".

Change-Id: Ia14441ba8d6f8140c21eeeb7ef0f0870bd72baf7
Reviewed-on: http://gerrit.aldebaran.lan/96229
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>