qi.type: Allows properties to have asynchronous accessors.
Might fix #44998.

This patch changes the accessors inside `PropertyImpl` into
asynchronous accessors that return futures. This allows clients to
construct properties directly with these asynchronous accessors,
avoiding the need, for instance, to wait in these accessors on some
other asynchronous function which result is needed to get or set the
value of the property.

The API of property types was already composed of member functions that
returned futures, so no change was introduced there. A set of new
constructors was added to property types.

This change enables us to correct some of the code of `ProxyProperty`
that previously needed to block on asynchronous remote access to a
property, that used to impose some non required thread blocking in
client code.

Change-Id: I43148d2f458b371ef482fba0885714b6ce452e4e
Reviewed-on: http://gerrit2.aldebaran.lan/215
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>