tests: Fixes use of object members concurrent with destruction.
In some tests, objects of the `Callsync` type could be used
during their destruction through their call operator. This would
lead to crashes or assertion failures.

This patch adds a `Trackable` in the type so that the use of the call
operator may be disabled at the beginning of destruction.