ka.zero: Adds the uninstantiable type `zero_t`.
Type with no instance.

Theoretically, it corresponds to the initial object in the category of
sets and functions, which is the empty set (i.e. the set with no element).
It is therefore the dual of `unit_t`.
It also happens to be (isomorphic to) the sum of no set, hence the
definition should be as an empty sum (but `boost::variant<>` and
`std::variant<>` are incomplete types, see TODO).

Warning: Abusing C++ type system is easy, and it is possible to form
  references or pointers to this uninstantiable type (e.g. with
  `(zero_t*)nullptr`). A good usage of this type of course excludes this
  kind of practices.

Warning: Copy constructor is not deactivated as a lot of existing code
  (`boost` libraries...) tends to instantiate types too greedily, and
  deactivation leads to a lot of artificial compilation problems. In a
  correct usage of this type, this lack should not be a problem since there
  is no value to copy in the first place.

meaning(zero_t) = 0

TODO: Make an alias of `ka::sum_t<>` when available.

Change-Id: I5e5ce5909592ec25c539d1d9e9446946aa55300f
Reviewed-on: http://gerrit.aldebaran.lan/97453
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>