libqi.application: fixes 'qi::Application' creation crash on protected path.
Fixes #45680

By default, 'qi::Application' tries to check if the path exists, but part of it
might be protected on some platforms, in which case a SIGABORT is triggered
because of an uncatched exception.

We made sure that each time an error occurs on path construction, including
access errors, an empty path is returned so we can track those errors and
avoid crashes.

Change-Id: I761bed51899cd9385a33968857a7abdb1bbc3723
Reviewed-on: http://gerrit2.aldebaran.lan/126
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: vincent.lartaud <vincent.lartaud@external.softbankrobotics.com>