qi.messaging: Forbids client TLS downgrade (spec:/sbre/framework/2020/a) (#45843)
To connect to a server through "tcps" (i.e. SSL/TLS), libqi 1.8.0 and 1.8.1
clients can accept to downgrade to SSL, if required by the server. The
motivation is retrocompatibility with older robots, but it appears this
downgrade is unnecessary.

This optional downgrade behavior has not been introduced by libqi 1.8.0 and
1.8.1, as they only raise the client "maximum" protocol from SSLv2/3 to TLS1.2.
Also, as this potential downgrade is entirely decided by the server, any NAOqi
gateway requiring TLS1.x cannot downgrade to SSL and will communicate with a
libqi 1.8.0 and 1.8.1 client with TLS1.x.

As a result, libqi 1.8.0 and 1.8.1 do not strictly weaken security, but they do
not make it as good as possible. Indeed, preventing client downgrade to TLS1.1,
i.e. making it use TLS 1.2 only, is enough for the targeted retrocompatibility.

A malicious server could still expose a libqi client 1.8.0 or 1.8.1 to a
"Man-in-the-Middle TLS Protocol Downgrade
Attack":https://www.praetorian.com/blog/man-in-the-middle-tls-ssl-protocol-downgrade-attack?edition=2019
justifying this patch.

Change-Id: I3a814dcb3fccd594eb5c83f9a5bdd126d496f133
Reviewed-on: http://gerrit2.aldebaran.lan/209
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>