qi.application: Sets and removes signal handlers during 'run' lifetime. #46111
There was a possible crash on application exit, because a callback that
occurred during the process exit tried to execute code using a program that
could be already closed.

Before this fix, handlers were removed during the process exit, causing a crash
when they were called outside of the library's lifetime.

The solution implemented by this fix is to manage handlers only in the run, to
make sure that every handler is removed before the program exits.  With this
fix, it's impossible that a handler still tries to access a program that has
already been closed.

Change-Id: If5466c73c201c08fccb246b510c0e92b5415d03f
Reviewed-on: http://gerrit2.aldebaran.lan/977
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: vincent.lartaud <vincent.lartaud@external.softbankrobotics.com>