libqi.application: fix stop handler inclusion in the signal sets list on run.
Signal sets were created using the global handler list insted of the local one.
But the local one added stop handler that had not been browsed to create its
signal sets.

This implied that the stop handler would never be called and the program would
not terminate correctly when a SIGTERM or a SIGINT is catched.

The solution is to browse the local list that has been completed instead of the
global one.

Change-Id: Ia2df4cd21c7b4dba94d660b1b4e9214b8482b535
Reviewed-on: http://gerrit2.aldebaran.lan/1103
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: vincent.lartaud <vincent.lartaud@external.softbankrobotics.com>