qi.application: Fixes high CPU consumption due to high-frequency `run` loop.
The loop in the `Application::run` function was causing an excessive CPU
consumption because of its short sleep duration.

This patch removes this loop and uses the blocking `run` function of the
`io_service` object instead. Doing this forces us to put the
`io_service` in a global static object.

Change-Id: Ia5a017203418cec544cb58e30b047424ad8df8f4
Reviewed-on: http://gerrit2.aldebaran.lan/1314
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: gerrit <gerrit@aldebaran-robotics.com>
Tested-by: gerrit <gerrit@aldebaran-robotics.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>