qi.os: Disables file access for machine-id retrieval on Android.
Fixes #42942.
Fixes #42899.
Fixes #40532.

This commit replaces any file access when trying to retrieve a
machine-id on Android, by the generation of a random id that persists
for the duration of the process. This is therefore not a real machine-id
but nonetheless closer to some than was the case before this commit,
since the function would always fail to access the file and would
therefore return a different id on each call.

It also rewrites the `getMachineId` function implementation to remove
the use of an `initialized` boolean and a mutex, and instead use a
`static` constant which is guaranteed to be initialized exactly once
during the process lifetime.

Change-Id: I40844872ff24aee38e3d8263bdb7b6e575caad19
Reviewed-on: http://gerrit2.aldebaran.lan/1491
Reviewed-by: gerrit <gerrit@aldebaran-robotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>