print: Fixes build error on newer versions of clang and its standard library.
Newer versions of the compiler and its standard library enforces that
`std::max_element` parameters' type must model `ForwardIterator`.

Unfortunately, `boost::iterator::transform_iterator` does not
implement `ForwardIterator` as it does
not always return a reference (as it returns the result of the function
which it was constructed from).

This patch rewrites the code so that we do not use
`boost::adaptors::transformed` anymore, but instead make `max_element`
use a custom comparison instead.

Change-Id: I6ec9fd6a01c66a4391b5656fe30b39d79f7ec2aa
Reviewed-on: http://gerrit2.aldebaran.lan/1568
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>