qi: Upgrades version of Boost from 1.64 to 1.77.
This patch ports the code to Boost 1.77, which comes with some
breaking API changes.

WARNING: This patch introduces an incompatibility with older versions of
Boost (Boost < 1.76).

Here is the detail of changes, ordered by library.

Boost.Asio
==========

The `get_io_service` function has been removed from many types and
replaced by an `executor` member function. Also, in order to
avoid too many changes in this patch, we've kept the use of some
deprecated types and functions.

Boost.Bind
==========

We've removed the `BOOST_BIND_GLOBAL_PLACEHOLDERS` definition passed
to the compiler and instead added it before each inclusion of
`boost/bind.hpp` in our public headers, so that our users code does not
break or generate lots of warnings because of the use of global
placeholders.

Boost.Locale
============

We've added a definition of `BOOST_LOCALE_HIDE_AUTO_PTR` which was added
in Boost 1.67 and that disables API using `std::auto_ptr`. This removes
unnecessary warnings about uses of the deprecated `auto_ptr` type on
recent standard libraries.

Boost.Program_options
=====================

The `boost::program_options::command_line_parser` type crashes when `argc`
is zero, so we had to add a test on that condition.

Boost.Predef
============

The `BOOST_OS_ANDROID` definition has been removed. We now use the
`ANDROID` definition instead.

Boost.Regex
===========

The dependency to the library has been removed, as Boost.Regex is now a
header-only library.

Boost.Timer
===========

`boost::timer` is now deprecated, but to avoid too much changes at
once, we keep using it. To avoid unnecessary noise, we've disabled a
warning about `boost/timer.hpp` being a deprecated header in the file
that includes it.

Change-Id: I9bc93c279fade2b28547e40db0db144168f605ca
Reviewed-on: http://gerrit2.aldebaran.lan/1592
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>