qi.messaging: Makes URI to URL conversion less strict.
The URL parsing function is a bit naive, and doesn't handle all cases
very well. This patch rewrites `qi::toUrl(Uri)` conversion function so
that it behaves like this:
  - the URL protocol is set to the URI scheme.
  - the URL host is set to the URI authority host iff it has one.
  - the URL port is set to the URI authority port iff it has one.

As such, it avoids using the parsing constructor of `qi::Url` that does
not handle corner cases coherently.
This makes it possible to convert URI that were previously not handled,
including qi relative URI such as 'qi:ServiceDirectory'.

Change-Id: I252538408d497fd2cf47018e328a2eb7e94b37d6
Reviewed-on: http://gerrit2.aldebaran.lan/1982
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Reviewed-by: philippe.martin <philippe.martin@softbankrobotics.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>