build: Rewrites project build definition. #SW-159
This patch rewrites the definition of the build of libqi in order to
remove the dependency to `qibuild` and `qi toolchains` and instead use
more standard tools and conventions.

It completely refactors the `CMakeLists.txt` files to use an approach
closer to what is referred to as "modern CMake". The project now requires
CMake 3.23 or greater.

The project declares the following targets:
- `qi::ka`: a header-only library.
- `qi::qi`: a shared library (compiled with PIC).
- `qi::testssl`: a static library that links with `qi::qi`.
- `qi::testsession`: a static library that links with `qi::qi`.

All targets are defined for:
- libqi as a CMake subproject (through `add_subdirectory`).
- a build directory of the project.
- an installation of the project.

The last two cases are implemented by generating package configuration
files (`X-config.cmake`) in the build directory and in the installation
directory.

Two package configuration files are created:
- `qi-config.cmake`
- `testsession-config.cmake`

Both declare the same targets, but with different namespaces.
The second one exists for compatibility with existing projects that use
`qibuild` and depend on a `testsession` package that used to be exported.

An additional `qimodule-config.cmake` file is generated but it does not
define targets. It instead defines a function for user projects to
declare qi modules.

The project defines 2 install components:
- `runtime`, which only installs files needed to run libqi, such as its shared
  libraries.
- `devel`, which installs files needed for development, such as headers
  and static libraries.

When installed, components use the following layout:
- `include/`: contains all public headers.
- `lib/`: contains shared and static libraries.
- `lib/cmake`: contains exported CMake package configuration files.

Tests are now declared to CMake using `add_test` and the CTest module is
used. This means that tests can now be executed with the `ctest` command
line tool.

This patch also adds the support of Conan, to download the project
dependencies and to create a Conan package of the project. Please refer
to the README for basic usage.