Fixes use of an iterator on a temporary vector.
The `ServiceInfo::endpoints` used to return a vector by const
reference. It now returns a vector by value. This causes some code that
did not store the result into a local variable to use iterators on the
temporary value.

This patch fixes the code by storing the vector in a local variable.