messaging: Fixes leak of socket objects.
Some code in `SessionService` sets a callback on a socket signal that
captures the socket by shared pointer, creating a cycle (the signal owns
the callback, the callback owns the socket and the socket owns the
signal).

Furthermore, the callback captures a shared pointer of the resulting
`SignalSubscriber` created from the connection, which itself owns the
callback.

This patch fixes both issues by:
  - capturing a weak pointer of the socket in the callback.
  - using a `SignalLink` instead of a `SignalSubscriber`.