messaging: Fixes leaks of RemoteObject.
A copy of the deleter passed to `makeDynamicAnyObject` is kept by
`boost::bind`. This means that a copy of the `shared_ptr` of
`RemoteObject` inside the deleter is also kept and the `RemoteObject`
never dies in some cases.

Furthermore, if when using a `DynamicObjectBuilder`, the `object(...)`
method is not called, the `DynamicObject` instantiated by the builder is
not destroyed.

This patch fixes these issues by:
  - using a deleter that moves the `onDelete` function to the
  `cleanupDynamicObject` function, ensuring that any data stored in
  the `onDelete` callback is destroyed once the deleter is executed.
  - using a boolean defined in `DynamicObjectBuilder` to know whether
  or not the `DynamicObject` is owned by the builder or if the ownership
  was moved to an `AnyObject`.