messaging.BoundObject: Fixes leak of tuple value.
When a call message has the `ReturnType` flag set, the function
parameters are in a tuple of dimension 2 containing first the
parameters, then the expected signature of the return value. Currently,
the tuple is decomposed by overriding the variable referencing it with
the reference to its first element, causing the reference to the tuple
to be lost. Unfortunately, this causes the tuple to not be properly
destroyed.

This patch fixes this issue by storing the value contained in the
message inside a local variable, and only referring to its subparts with
non-owning references.