type: Fixes wrong unwrapping of `Any{Value,Reference}` in optional interface.
Expected behavior
=================

The implementation of the qi type system assumes the following:

$$
With kind: Type → Kind,
     Dynamic: Kind → Kind,
     ∀ k: Kind
        Dynamic(k) ≠ k:
  ∀ t: Type, Any ∈ {AnyReference, AnyValue, AutoAnyReference}
    kind(Any(t)) = Dynamic(kind(t))
$$

And in particular:

$$
∀ t: Type, Any ∈ {AnyReference, AnyValue}
  kind(AutoAnyReference(Any(t))) = Dynamic(kind(Any(t)))
                                 = Dynamic(Dynamic(kind(t)))
$$

Real behavior
=============

In the optional type interface implementation, the `value(void*)` method
returns a reference to the value inside the optional (or to `void` if
empty). To perform type deduction, the `AutoAnyReference` type was used
to construct this reference. However, `AutoAnyReference` has a
constructor that takes a `const AnyReferenceBase&` value
(`AnyReferenceBase` being the base type from both `AnyReference` and
`AnyValue`). This constructor creates the reference from the passed
`AnyReference/AnyValue` by effectively unwrapping it. Therefore, we
ended up with:

$$
∀ t: Type, Any ∈ {AnyReference, AnyValue}
  kind(AutoAnyReference(Any(t))) = Dynamic(kind(t))
                                 ≠ Dynamic(Dynamic(kind(t)))
$$

This caused an issue in the interface implementation in the case of
`boost::optional<AnyValue>`. Instead of returning an `AnyReference` to
the `AnyValue` contained in the optional, the `value(void*)` method
returned an `AnyReference` to the value contained in the `AnyValue`
contained in the optional. This caused inconsistency between the
signature of the type of the value contained in the optional
(effectively `Dynamic(T)`) and the signature of the type of the value
obtained when unwrapping the optional (effectively `T`).

Fix
===

This patch fixes this issue by using `AnyReference::from` instead of
`AutoAnyReference`, which does no value unwrapping.